modified way of random throws
now the fun begins:
I renamed 'lot' to 'bucket' and we have an aim now:
each bucket must be filled with a given amount of random results.

See how interesting the histogram starts to look
with just a few throws in each bucket.
It is quite ok, as the number of throws increases,
the uniformity improves.